test: Move `test-uuid` module to global libs
