+++ node/packages/aws-lambda-sdk/test/lib/basename.js
@@ -1,5 +1,5 @@
 'use strict';
 
-const testUid = require('./test-uid');
+const testUid = require('../../../../test/lib/test-uid');
 
 module.exports = `test-sdk-${testUid}`;
